Title: 소괄호 매칭 수 최대화하기, Time: 1012ms, Memory: 4MB, Status: Time Limit Exceed - CodeTree

+++ 240709/소괄호 매칭 수 최대화하기/maximize-the-number-of-parenthesis-matches.cpp
@@ -5,8 +5,8 @@
 
 using namespace std;
 
-int calcPoint(string str) {
-    int point = 0;
+long long calcPoint(string str) {
+    long long point = 0;
     vector <int> count_arr;
 
     int prev_cnt = 0;
